sba/requirements.txt
django<5
django-dotenv<2
django-livereload-server<1
django-widget-tweaks<2
requests<3
whitenoise<7
psycopg2-binary<3